+++ src/main/java/insane96mcp/iguanatweaksexpanded/IguanaTweaksExpanded.java
@@ -8,6 +8,7 @@ import insane96mcp.iguanatweaksexpanded.data.generator.client.ITEBlockStatesProv
 import insane96mcp.iguanatweaksexpanded.data.generator.client.ITEItemModelsProvider;
 import insane96mcp.iguanatweaksexpanded.module.combat.fletching.Fletching;
 import insane96mcp.iguanatweaksexpanded.module.combat.fletching.dispenser.ITEArrowDispenseBehaviour;
+import insane96mcp.iguanatweaksexpanded.module.items.recallidol.Recall;
 import insane96mcp.iguanatweaksexpanded.network.NetworkHandler;
 import insane96mcp.iguanatweaksexpanded.setup.ITECommonConfig;
 import insane96mcp.iguanatweaksexpanded.setup.ITEPackSource;
@@ -37,6 +38,7 @@ import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
 import net.minecraftforge.fml.event.lifecycle.FMLClientSetupEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
+import net.minecraftforge.fml.event.lifecycle.FMLLoadCompleteEvent;
 import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 import net.minecraftforge.fml.loading.FMLLoader;
 import org.apache.logging.log4j.LogManager;
@@ -72,6 +74,7 @@ public class IguanaTweaksExpanded
             modEventBus.addListener(ClientSetup::registerParticleFactories);
         }
         modEventBus.addListener(this::commonSetup);
+        modEventBus.addListener(this::loadComplete);
         modEventBus.addListener(this::clientSetup);
         modEventBus.addListener(this::gatherData);
         modEventBus.addListener(this::addPackFinders);
@@ -83,6 +86,7 @@ public class IguanaTweaksExpanded
 
     private void commonSetup(final FMLCommonSetupEvent event) {
         NetworkHandler.init();
+        Recall.onLoadComplete();
 
         event.enqueueWork(() -> {
             DispenserBlock.registerBehavior(Fletching.QUARTZ_ARROW_ITEM.get(), new ITEArrowDispenseBehaviour());
@@ -142,4 +146,11 @@ public class IguanaTweaksExpanded
             event.getServer().reloadResources(list.stream().map(Pack::getId).collect(Collectors.toList()));
     }
 
+    private void loadComplete(final FMLLoadCompleteEvent event)
+    {
+        event.enqueueWork(() ->
+        {
+            Recall.onLoadComplete();
+        });
+    }
 }